+++ src/main/java/com/github/theredbrain/spellengineextension/SpellEngineExtension.java
@@ -9,6 +9,8 @@ import com.github.theredbrain.spellengineextension.compat.StaminaAttributesCompa
 import com.github.theredbrain.spellengineextension.component.type.HasConditionalSpellContainerComponent;
 import com.github.theredbrain.spellengineextension.config.ServerConfig;
 import com.github.theredbrain.spellengineextension.entity.player.DuckPlayerEntityMixin;
+import com.github.theredbrain.spellengineextension.predicate.item.SpellContainersPredicate;
+import com.github.theredbrain.spellengineextension.registry.DataComponentPredicateRegistry;
 import com.github.theredbrain.spellengineextension.registry.ItemComponentRegistry;
 import com.github.theredbrain.spellengineextension.registry.SpellSchoolRegistry;
 import com.github.theredbrain.spellengineextension.spell_engine.DuckSpellActiveCastMixin;
@@ -23,6 +25,7 @@ import net.minecraft.entity.attribute.EntityAttributes;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.entity.player.PlayerInventory;
 import net.minecraft.item.ItemStack;
+import net.minecraft.predicate.item.ItemSubPredicate;
 import net.minecraft.registry.entry.RegistryEntry;
 import net.minecraft.registry.tag.TagKey;
 import net.minecraft.util.Identifier;
@@ -57,6 +60,8 @@ public class SpellEngineExtension implements ModInitializer {
 
 	public static ComponentType<HasConditionalSpellContainerComponent> HAS_CONDITIONAL_SPELL_CONTAINER;
 	public static ComponentType<Identifier> PROXY_POOL;
+
+	public static ItemSubPredicate.Type<SpellContainersPredicate> SPELL_CONTAINER_PREDICATE;
 
 	public static final TagKey<Spell> CAN_BE_IN_USE_ITEM_SPELL_HOTBAR_SLOT = TagKey.of(SpellRegistry.KEY, SpellEngineExtension.identifier("can_be_in_use_item_spell_hotbar_slot"));
 	public static final TagKey<Spell> ENABLES_MOVEMENT_LOCKING_DURING_CASTING = TagKey.of(SpellRegistry.KEY, SpellEngineExtension.identifier("enables_movement_locking_during_casting"));
@@ -148,6 +153,7 @@ public class SpellEngineExtension implements ModInitializer {
 		LOGGER.info("Spell Engine was extended!");
 		SERVER_CONFIG = ConfigApiJava.registerAndLoadConfig(ServerConfig::new, RegisterType.BOTH);
 
+		DataComponentPredicateRegistry.init();
 		ItemComponentRegistry.init();
 		SpellSchoolRegistry.init();